syllabus

+++ server/controllers/admin.controller.js
@@ -2192,6 +2192,65 @@ exports.getAssignedTasks = async (req, res) => {
 };
 
 
+// exports.createOrUpdateSyllabus = async (req, res) => {
+//   try {
+//     const loggedInId = req.user && req.user.id;
+//     if (!loggedInId) {
+//       return res.status(401).json({ message: 'Unauthorized, only loggedin users can access this.' })
+//     };
+
+//     const loggedInUser = await User.findById(loggedInId);
+//     if (!loggedInUser || loggedInUser.role !== 'admin') {
+//       return res.status(403).json({
+//         message: 'Access denied, only admin have access.'
+//       })
+//     };
+
+//     const { className, subject, description } = req.body;
+//     if (!className || !subject || !description) { return res.status(400).json({ message: "Please provide class, subject, description to upload syllabus." }) }
+
+//     let subjectIs = subject.toLowerCase();
+
+//     const school = await School.findOne({ userId: loggedInId });
+//     if (!school) { return res.status(404).json({ message: "You are not associated with any school." }) }
+
+//     let uploadedPhotoUrl = '';
+//     if (req.file) {
+//       try {
+//         const [photoUrl] = await uploadImage(req.file);
+//         uploadedPhotoUrl = photoUrl;
+//       } catch (error) {
+//         return res.status(500).json({ message: 'Failed to upload photo.', error: error.message });
+//       }
+//     }
+
+//     let existingSyllabus = await Syllabus.findOne({ schoolId: school._id, class: className, subject: subjectIs });
+
+//     if (existingSyllabus) {
+//       if (existingSyllabus.syllabus) {
+//         await deleteImage(existingSyllabus.syllabus);
+//         existingSyllabus.syllabus = uploadedPhotoUrl;
+//       }
+//       existingSyllabus.class = className;
+//       existingSyllabus.subject = subjectIs;
+//       existingSyllabus.description = description;
+
+//       await existingSyllabus.save();
+
+//       return res.status(200).json({ message: 'Syllabus updated successfully.', existingSyllabus });
+//     }
+//     else {
+//       const newSyllabus = new Syllabus({ schoolId: school._id, class: className, subject: subjectIs, description, syllabus: uploadedPhotoUrl, });
+//       await newSyllabus.save();
+
+//       return res.status(201).json({ message: 'Syllabus created successfully.', newSyllabus, });
+//     }
+//   }
+//   catch (err) {
+//     res.status(500).json({ message: 'Internal server error.', error: err.message, });
+//   }
+// };
+
 exports.createOrUpdateSyllabus = async (req, res) => {
   try {
     const loggedInId = req.user && req.user.id;
@@ -2206,15 +2265,17 @@ exports.createOrUpdateSyllabus = async (req, res) => {
       })
     };
 
-    const { className, subject, description } = req.body;
-    if (!className || !subject || !description) { return res.status(400).json({ message: "Please provide class, subject, description to upload syllabus." }) }
+    let uploadedPhotoUrl, existingSyllabus, subjectIs;
 
-    let subjectIs = subject.toLowerCase();
+    const { className, subject, description } = req.body;
+
+    const { id } = req.params;
+
+    if (subject) subjectIs = subject.toLowerCase();
 
     const school = await School.findOne({ userId: loggedInId });
     if (!school) { return res.status(404).json({ message: "You are not associated with any school." }) }
 
-    let uploadedPhotoUrl = '';
     if (req.file) {
       try {
         const [photoUrl] = await uploadImage(req.file);
@@ -2224,22 +2285,27 @@ exports.createOrUpdateSyllabus = async (req, res) => {
       }
     }
 
-    let existingSyllabus = await Syllabus.findOne({ schoolId: school._id, class: className, subject: subjectIs });
+    if (id) {
+      if (!className && !subject && !description && !req.file) { return res.status(400).json({ message: "Please provide atleast one valid field to update syllabus." }) }
+      existingSyllabus = await Syllabus.findOne({ schoolId: school._id, _id: id });
+    }
 
     if (existingSyllabus) {
-      if (existingSyllabus.syllabus) {
+      if (existingSyllabus.syllabus && req.file) {
         await deleteImage(existingSyllabus.syllabus);
         existingSyllabus.syllabus = uploadedPhotoUrl;
       }
-      existingSyllabus.class = className;
-      existingSyllabus.subject = subjectIs;
-      existingSyllabus.description = description;
+      if (className) existingSyllabus.class = className;
+      if (subject) existingSyllabus.subject = subjectIs;
+      if (description) existingSyllabus.description = description;
 
       await existingSyllabus.save();
 
       return res.status(200).json({ message: 'Syllabus updated successfully.', existingSyllabus });
     }
     else {
+      if (!className || !subject || !description || !req.file) { return res.status(400).json({ message: "Please provide class, subject, description and syllabus file to upload." }) }
+
       const newSyllabus = new Syllabus({ schoolId: school._id, class: className, subject: subjectIs, description, syllabus: uploadedPhotoUrl, });
       await newSyllabus.save();
 
@@ -2842,7 +2908,7 @@ exports.issueAndReturnBook = async (req, res) => {
 
     if (status == 'accepted' || status == 'rejected' || status == 'requested') {
       bookRequest.status = status,
-      bookRequest.borrowedOn = null,
+        bookRequest.borrowedOn = null,
         bookRequest.dueOn = null,
         bookRequest.returnedOn = null,
         bookRequest.fine = 0,
